Second go at fixing charm migration
